almost done

+++ src/redux/pasteSlice.js
@@ -13,19 +13,54 @@ export const pasteSlice = createSlice({
   reducers: {
     addToPastes: (state, action) => {
       const paste = action.payload;
+
+      // adding check -> Paste allready exist 
+      // ✅ Check if paste already exists (based on `_id`)
+      const isExist = state.pastes.some(p => p._id === paste._id);
+
+      if (isExist) {
+        toast.error("Paste already exists!")
+        return;
+      }
+
+      // ❌ If not exists, then add
       state.pastes.push(paste);
       localStorage.setItem("pastes",
         JSON.stringify(state.pastes));
         toast("Paste Created Sucessfully")
     },
     updateToPastes: (state, action) => {
-      
+      const paste = action.payload;
+      const index = state.pastes.findIndex((item) => 
+        item._id === paste._id);
+
+      if (index >= 0) {
+        state.pastes[index] = paste;
+
+        localStorage.setItem("pastes", JSON.stringify(state.pastes));
+
+        toast.success("Paste Updated")
+      }
     },
     resetAllPastes: (state, action) => {
+      state.pastes = [];
 
+      localStorage.removeItem("pastes");
     },
     removeFromPastes: (state, action) => {
+      const pasteId = action.payload;
 
+      console.log(pasteId);
+      const index = state.pastes.findIndex((item) =>
+      item._id === pasteId);
+
+      if(index >= 0) {
+        state.pastes.splice(index, 1);
+
+        localStorage.setItem("pastes", JSON.stringify(state.pastes));
+
+        toast.success("paste deleted");
+      }
     },
   },
 })
